support partitioned topic producers in client crate
