+++ danube-client/src/lib.rs
@@ -14,8 +14,12 @@ pub mod errors;
 mod producer;
 pub use producer::{Producer, ProducerBuilder, ProducerOptions};
 
+mod topic_producer;
+
 mod consumer;
 pub use consumer::{Consumer, ConsumerBuilder, ConsumerOptions, SubType};
+
+mod message_router;
 
 mod schema;
 pub use schema::{Schema, SchemaType};
