+++ danube-client/src/message_router.rs
@@ -1,0 +1,12 @@
+/// Represents a Producer
+#[derive(Debug)]
+pub(crate) struct MessageRouter {}
+
+impl MessageRouter {
+    pub(crate) fn new() -> Self {
+        MessageRouter {}
+    }
+    pub(crate) fn get_partition(&self) -> usize {
+        return 1;
+    }
+}
